refactored TableColumnCsvExtractor to the csv_extractors file. removing sample_csv_data_loader.py simplified sample_data_loader.py Finally, modified sample data so that popular table shows by default again.
